Kernal patch
Kernal.php was missing from the project for some reason. It was supposed to be in the base laravel app.

I'm still unsure if there was an error in the app generation, or if it was removed during development.

I've implemented the Kernal.php file to ammend this oversight as well as associated middleware so I can continue development on my other branch.